warning fixes unsigned->signed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@29373 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/mac/carbon/checklst.cpp
@@ -74,7 +74,7 @@ static pascal  void DataBrowserItemNotificationProc(ControlRef browser, DataBrow
     {
         wxCheckListBox* list = wxDynamicCast( (wxObject*) ref , wxCheckListBox ) ;
         size_t i = itemID - 1 ;
-        if (i >= 0 && i < list->GetCount() )
+        if (i >= 0 && i < (size_t) list->GetCount() )
         {
             bool trigger = false ;
             wxCommandEvent event(
@@ -133,7 +133,7 @@ static pascal OSStatus ListBoxGetSetItemData(ControlRef browser,
     		    {
     		        wxCheckListBox* list = wxDynamicCast( (wxObject*) ref , wxCheckListBox ) ;
                     size_t i = itemID - 1 ;
-                    if (i >= 0 && i < list->GetCount() )
+                    if (i >= 0 && i < (size_t) list->GetCount() )
 		            {
 		                wxMacCFStringHolder cf( list->GetString(i) , list->GetFont().GetEncoding() ) ;
 		                verify_noerr( ::SetDataBrowserItemDataText( itemData , cf ) ) ;
@@ -149,7 +149,7 @@ static pascal OSStatus ListBoxGetSetItemData(ControlRef browser,
     		    {
     		        wxCheckListBox* list = wxDynamicCast( (wxObject*) ref , wxCheckListBox ) ;
                     size_t i = itemID - 1 ;
-                    if (i >= 0 && i < list->GetCount() )
+                    if (i >= 0 && i < (size_t) list->GetCount() )
 		            {
 		                verify_noerr( ::SetDataBrowserItemDataButtonValue( itemData , list->IsChecked( i ) ? kThemeButtonOn : kThemeButtonOff ) ) ;
 		                err = noErr ;
@@ -178,7 +178,7 @@ static pascal OSStatus ListBoxGetSetItemData(ControlRef browser,
     		    {
     		        wxCheckListBox* list = wxDynamicCast( (wxObject*) ref , wxCheckListBox ) ;
                     size_t i = itemID - 1 ;
-                    if (i >= 0 && i < list->GetCount() )
+                    if (i >= 0 && i < (size_t) list->GetCount() )
 		            {
 		                // we have to change this behind the back, since Check() would be triggering another update round
                         bool newVal = !list->IsChecked( i ) ;
